Revert "fixed sales" (#81)
* hot fix

* fixed delete bundle spec

* active user deletion

* fixed message

* hot fix

+++ app/src/main/java/it/gym/facade/UserFacade.java
@@ -1,10 +1,15 @@
 package it.gym.facade;
 
+import it.gym.exception.BadRequestException;
 import it.gym.exception.NotFoundException;
 import it.gym.model.AUser;
 import it.gym.model.VerificationToken;
 import it.gym.service.UserService;
 import it.gym.service.VerificationTokenService;
+import org.hibernate.JDBCException;
+import org.hibernate.exception.ConstraintViolationException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.data.domain.Page;
@@ -17,6 +22,8 @@ import javax.transaction.Transactional;
 @Component
 @Transactional
 public class UserFacade {
+
+    private final Logger logger = LoggerFactory.getLogger(getClass());
 
     @Autowired
     @Qualifier("verificationTokenService")
@@ -31,7 +38,13 @@ public class UserFacade {
             VerificationToken vtk = this.tokenService.findByUser(user);
             this.tokenService.delete(vtk);
         }
-        this.service.deleteById(id);
+
+        if (!user.isActive()) {
+            this.service.deleteById(id);
+        }
+        else {
+            throw new BadRequestException("Impossibile eliminare un utente con delle transazioni attive");
+        }
         return user;
     }
